feature: make independent modules from single file
